+++ src/community/community_misc.c
@@ -270,6 +270,11 @@ int igraph_reindex_membership(igraph_vector_t *membership,
     for (i = 0; i < n; i++) {
         long int c = (long int)VECTOR(*membership)[i];
 
+        if (c < 0) {
+            IGRAPH_ERRORF("Membership indices should non-negative. "
+            "Found member of cluster %ld.", IGRAPH_EINVAL, c);
+        }
+
         if (c >= n) {
             IGRAPH_ERRORF("Membership indices should be less than total number of vertices. "
             "Found member of cluster %ld, but only %ld vertices.", IGRAPH_EINVAL, c, n);